mito-ai: refactor websockets
