+++ mito-ai/mito_ai/completions/completion_handlers/agent_auto_error_fixup_handler.py
@@ -2,13 +2,13 @@
 # Distributed under the terms of the GNU Affero General Public License v3.0 License.
 
 from openai.types.chat import ChatCompletionMessageParam
-from mito_ai.models import AgentResponse, AgentSmartDebugMetadata, MessageType, ResponseFormatInfo
-from mito_ai.prompt_builders.agent_smart_debug_prompt import create_agent_smart_debug_prompt
-from mito_ai.providers import OpenAIProvider
-from mito_ai.message_history import GlobalMessageHistory
-from mito_ai.completion_handlers.completion_handler import CompletionHandler
-from mito_ai.completion_handlers.open_ai_models import MESSAGE_TYPE_TO_MODEL
-from mito_ai.completion_handlers.utils import append_agent_system_message
+from mito_ai.completions.models import AgentResponse, AgentSmartDebugMetadata, MessageType, ResponseFormatInfo
+from mito_ai.completions.prompt_builders.agent_smart_debug_prompt import create_agent_smart_debug_prompt
+from mito_ai.completions.providers import OpenAIProvider
+from mito_ai.completions.message_history import GlobalMessageHistory
+from mito_ai.completions.completion_handlers.completion_handler import CompletionHandler
+from mito_ai.completions.completion_handlers.open_ai_models import MESSAGE_TYPE_TO_MODEL
+from mito_ai.completions.completion_handlers.utils import append_agent_system_message
 
 
 __all__ = ["get_agent_auto_error_fixup_completion"]
