+++ mito-ai/mito_ai/completions/completion_handlers/agent_execution_handler.py
@@ -3,13 +3,13 @@
 
 from typing import List, Literal, Union
 from openai.types.chat import ChatCompletionMessageParam
-from mito_ai.models import AgentExecutionMetadata, MessageType, ResponseFormatInfo, AgentResponse
-from mito_ai.prompt_builders.agent_execution_prompt import create_agent_execution_prompt
-from mito_ai.providers import OpenAIProvider
-from mito_ai.message_history import GlobalMessageHistory
-from mito_ai.completion_handlers.completion_handler import CompletionHandler
-from mito_ai.completion_handlers.open_ai_models import MESSAGE_TYPE_TO_MODEL
-from mito_ai.completion_handlers.utils import append_agent_system_message, create_ai_optimized_message
+from mito_ai.completions.models import AgentExecutionMetadata, MessageType, ResponseFormatInfo, AgentResponse
+from mito_ai.completions.prompt_builders.agent_execution_prompt import create_agent_execution_prompt
+from mito_ai.completions.providers import OpenAIProvider
+from mito_ai.completions.message_history import GlobalMessageHistory
+from mito_ai.completions.completion_handlers.completion_handler import CompletionHandler
+from mito_ai.completions.completion_handlers.open_ai_models import MESSAGE_TYPE_TO_MODEL
+from mito_ai.completions.completion_handlers.utils import append_agent_system_message, create_ai_optimized_message
 
 __all__ = ["get_agent_execution_completion"]
