+++ mito-ai/mito_ai/completions/message_history.py
@@ -9,10 +9,10 @@ from threading import Lock
 from typing import Dict, List, Optional
 
 from openai.types.chat import ChatCompletionMessageParam
-from mito_ai.completion_handlers.open_ai_models import MESSAGE_TYPE_TO_MODEL
-from mito_ai.models import CompletionRequest, ChatThreadMetadata, MessageType, ThreadID
-from mito_ai.prompt_builders.chat_name_prompt import create_chat_name_prompt
-from mito_ai.providers import OpenAIProvider
+from mito_ai.completions.completion_handlers.open_ai_models import MESSAGE_TYPE_TO_MODEL
+from mito_ai.completions.models import CompletionRequest, ChatThreadMetadata, MessageType, ThreadID
+from mito_ai.completions.prompt_builders.chat_name_prompt import create_chat_name_prompt
+from mito_ai.completions.providers import OpenAIProvider
 from mito_ai.utils.schema import MITO_FOLDER
 from mito_ai.utils.message_history_utils import trim_old_messages
