+++ mito-ai/mito_ai/completions/prompt_builders/chat_prompt.py
@@ -2,7 +2,7 @@
 # Distributed under the terms of the GNU Affero General Public License v3.0 License.
 
 from typing import List
-from mito_ai.prompt_builders.prompt_constants import (
+from mito_ai.completions.prompt_builders.prompt_constants import (
     ACTIVE_CELL_ID_SECTION_HEADING,
     FILES_SECTION_HEADING,
     VARIABLES_SECTION_HEADING,
