+++ mito-ai/mito_ai/completions/prompt_builders/explain_code_prompt.py
@@ -1,7 +1,7 @@
 # Copyright (c) Saga Inc.
 # Distributed under the terms of the GNU Affero General Public License v3.0 License.
 
-from mito_ai.prompt_builders.prompt_constants import CODE_SECTION_HEADING
+from mito_ai.completions.prompt_builders.prompt_constants import CODE_SECTION_HEADING
 
 def create_explain_code_prompt(active_cell_code: str) -> str:
     prompt = f"""Explain the code in the active code cell to me like I have a basic understanding of Python. Don't explain each line, but instead explain the overall logic of the code.
