+++ mito-ai/mito_ai/utils/message_history_utils.py
@@ -4,7 +4,7 @@
 import re
 from typing import List
 from openai.types.chat import ChatCompletionMessageParam
-from mito_ai.prompt_builders.prompt_constants import (
+from mito_ai.completions.prompt_builders.prompt_constants import (
     ACTIVE_CELL_ID_SECTION_HEADING,
     ACTIVE_CELL_OUTPUT_SECTION_HEADING,
     GET_CELL_OUTPUT_TOOL_RESPONSE_SECTION_HEADING,
